requirements.txt
tqdm
beautifulsoup4
requests
python-dotenv
pyyaml
datasets
huggingface_hub
pandas
pillow